LPS-55590 Copy EditUserAction, no logic changes
